% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{anovaTest}
\alias{anovaTest}
\title{Computes the ANOVA-Test for clustered data}
\usage{
anovaTest(data, cont, theta = NULL, psi = NULL, alpha = 0.05, type = NULL)
}
\arguments{
\item{data}{The data, provided as a list of lists}

\item{cont}{A contrast matrix for the relative effects}

\item{theta}{A Vector for the group weights, defaults to unweighted estimator}

\item{psi}{A list of vectors with the cluster weights, defaults to unweighted estimator}

\item{alpha}{The significance level, defaults to 0.05}

\item{type}{A string indicating whether weighted or unweighted estimator should be used. Only if psi is not provided}
}
\value{
A list containing the value of the test statistic, the degrees of freedom, the p-value and the test decision
}
\description{
Computes the ANOVA-Test for clustered data
}
