% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{big4}
\alias{big4}
\title{Rank data: big4}
\format{
A list containing:
\describe{
  \item{data}{A matrix of size 21*8 containing the 21 Premier League seasons. Each row corresponding to one ranking (ranking representation).

             The ranking representation r=(r_1,...,r_m) contains the  ranks assigned to the objects, and means that the ith object is in r_ith position.

              For example, if the ranking representation of a rank is (4,3,1,2,5), it means that judge ranks the first object in 4th position, second object in 3rd position, ...}
  \item{frequency}{matrix of size 21*9. Each row corresponds to one of the 21 different observed rankings, and the last column contains the observation frequency.}
  \item{m}{the size of the rankings (m=c(4,4) ).}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Premier_League}

\url{https://www.uefa.com/memberassociations/uefarankings/club/}
}
\description{
This dataset is composed of the rankings (in ranking notation) of the "Big Four" English football teams (A: Manchester, B: Liverpool, C: Arsenal, D: Chelsea) to the English Championship (Premier League) and according to the UEFA coefficients (statistics used in Europe for ranking and seeding teams in international competitions), from 1993 to 2013.

In 2000-2001, Arsenal and Chelsea had the same UEFA coefficient and then are tied. UEFA ranking is (1, 4, 2, 2) for 2000-2001, what means that Manchester United is the first, Liverpool is the last, and the two intermediate positions are for Arsenal and Chelsea in an unknown order.

In 2009-2010, Liverpool and Arsenal have also the same UEFA coefficient, the ranking is  (1, 2, 2, 4).
}
\examples{
data(big4)

}
\seealso{
Other datasets: 
\code{\link{APA}},
\code{\link{eurovision}},
\code{\link{quiz}},
\code{\link{sports}},
\code{\link{words}}
}
\concept{datasets}
\keyword{datasets}
