% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{clusterTest}
\alias{clusterTest}
\title{Computes comprehensive analysis of the cluster data, containing descriptive statistics, confidence intervals and different test procedures.}
\usage{
clusterTest(
  data,
  p_null = 0.5,
  contrast = NULL,
  normal = FALSE,
  theta = NULL,
  psi = NULL,
  alpha = 0.05,
  type = NULL
)
}
\arguments{
\item{data}{The data, provided as a list of lists}

\item{p_null}{A vector or scalar of relative effects under the null hypothesis, defaults to 0.5}

\item{contrast}{A contrast statement, either given as a numeric matrix or a character string (e.g. "Dunnett", "Tukey", "Sequen", "AVE", "Changepoint", "Williams", "Marcus", "McDermott", "UmbrellaWilliams", "GrandMean"). Defaults to the "GrandMean".}

\item{normal}{Logical, whether normal approximation should be used or not, defaults to FALSE}

\item{theta}{A Vector for the group weights, defaults to unweighted estimator}

\item{psi}{A list of vectors with the cluster weights, defaults to unweighted estimator}

\item{alpha}{The significance level, defaults to 0.05}

\item{type}{A string indicating whether weighted or unweighted estimator should be used. Only if psi is not provided}
}
\description{
Computes comprehensive analysis of the cluster data, containing descriptive statistics, confidence intervals and different test procedures.
}
