% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankDistance.R
\name{distCayley}
\alias{distCayley}
\title{Cayley distance between two ranks}
\usage{
distCayley(x, y)
}
\arguments{
\item{x, y}{two ranks of size m.}
}
\value{
the Cayley distance between x and y.
}
\description{
The Cayley distance between two ranks x and y is the minimum number of transpositions required to transform the ranking x into y.
}
\examples{
x <- 1:5
y <- c(2, 3, 1, 4, 5)
distCayley(x, y)

}
\seealso{
Other distance: 
\code{\link{distHamming}()},
\code{\link{distKendall}()},
\code{\link{distSpearman}()}
}
\author{
Julien Jacques
}
\concept{distance}
