% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankDistance.R
\name{distHamming}
\alias{distHamming}
\title{Hamming distance between two ranks}
\usage{
distHamming(x, y)
}
\arguments{
\item{x, y}{two ranks of size m.}
}
\value{
an integer, the Hamming distance between x and y.
}
\description{
The Hamming distance between two ranks x and y is the number of difference between the two ranks.
For example, the Hamming's distance between x=(1,4,2,5,3) and y=(1,3,4,5,2) is 3 because, only 1 and 5 have the same place in both ranks.
}
\examples{
x <- 1:5
y <- c(2, 3, 1, 4, 5)
distHamming(x, y)

}
\seealso{
Other distance: 
\code{\link{distCayley}()},
\code{\link{distKendall}()},
\code{\link{distSpearman}()}
}
\author{
Julien Jacques
}
\concept{distance}
