% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankDistance.R
\name{distSpearman}
\alias{distSpearman}
\title{Spearman distance between two ranks}
\usage{
distSpearman(x, y)
}
\arguments{
\item{x, y}{two ranks of size m.}
}
\value{
an integer, the Spearman distance between x and y.
}
\description{
The Spearman distance is the square of Euclidean distance between two rank vector.
}
\examples{
x <- 1:5
y <- c(2, 3, 1, 4, 5)
distSpearman(x, y)

}
\seealso{
Other distance: 
\code{\link{distCayley}()},
\code{\link{distHamming}()},
\code{\link{distKendall}()}
}
\author{
Julien Jacques
}
\concept{distance}
