% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{f_theta}
\alias{f_theta}
\title{Empirical Reference Distribution Function with group weights theta and cluster weights psi}
\usage{
f_theta(x, data, theta = NULL, psi = NULL)
}
\arguments{
\item{x}{A scalar or vector to be checked as F(x)}

\item{data}{The data, provided as a list of lists}

\item{theta}{A vector of group weights, defaults to unweighted estimator}

\item{psi}{A list of vector-weights, defaults to unweighted estimator}
}
\value{
Value(s) of the empirical distribution function
}
\description{
Empirical Reference Distribution Function with group weights theta and cluster weights psi
}
