% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{g}
\alias{g}
\title{Inflation-term}
\usage{
g(data, type = "unweighted")
}
\arguments{
\item{data}{The data, provided as a list of lists}

\item{type}{A string declaring whether a weighted or unweighted estimator will be used, defaults to "unweighted"}
}
\value{
Scalar of inflation
}
\description{
Inflation-term
}
