% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{h_0_f}
\alias{h_0_f}
\title{Data-generating function under the null hypothesis}
\usage{
h_0_f(
  n = rep(5, 5),
  m = NULL,
  dist = "norm",
  corstruct = "independent",
  rho = NULL,
  rho.max = NULL,
  params = list(mean = 0, sd = 1)
)
}
\arguments{
\item{n}{Vector of sample sizes}

\item{m}{Vector of cluster sizes, defaults to sample size}

\item{dist}{Distribution under the null, defaults to normal}

\item{corstruct}{Type of correlation (string), defaults to "independent". Alternatives are "exchangeable" and "wild"}

\item{rho}{Fixed correlation, defaults to NULL}

\item{rho.max}{In case of no fixed correlation, specifies the maximal correlation, defaults to NULL}

\item{params}{List of params necessary for the distribution specified in dist, defaults to standard normal}
}
\value{
Returns a list of lists with the data
}
\description{
Data-generating function under the null hypothesis
}
