% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{max_T}
\alias{max_T}
\title{Computes the ANOVA-Test for clustered data}
\usage{
max_T(
  n,
  data,
  p_null = 0.5,
  cont,
  normal = FALSE,
  theta = NULL,
  psi = NULL,
  alpha
)
}
\arguments{
\item{n}{A vector containing the sample sizes}

\item{data}{The data, provided as a list of lists}

\item{p_null}{A vector or scalar of relative effects under the null hypothesis, defaults to 0.5}

\item{cont}{A contrast matrix for the relative effects}

\item{normal}{Logical, whether normal approximation should be used or not, defaults to FALSE}

\item{theta}{A Vector for the group weights, defaults to unweighted estimator}

\item{psi}{A list of vectors with the cluster weights, defaults to unweighted estimator}

\item{alpha}{The significance level, defaults to 0.05}
}
\value{
A list containing the value of the test statistic, the degrees of freedom and the test decision
}
\description{
Computes the ANOVA-Test for clustered data
}
