% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{mctp}
\alias{mctp}
\title{Computes the MCTP for clustered data}
\usage{
mctp(
  data,
  p_null = 0.5,
  cont,
  normal = FALSE,
  theta = NULL,
  psi = NULL,
  alpha,
  type = NULL
)
}
\arguments{
\item{data}{The data, provided as a list of lists}

\item{p_null}{A vector or scalar of relative effects under the null hypothesis, defaults to 0.5}

\item{cont}{A contrast matrix for the relative effects}

\item{normal}{Logical, whether normal approximation should be used or not, defaults to FALSE}

\item{theta}{A Vector for the group weights, defaults to unweighted estimator}

\item{psi}{A list of vectors with the cluster weights, defaults to unweighted estimator}

\item{alpha}{The significance level, defaults to 0.05}

\item{type}{A string indicating whether weighted or unweighted estimator should be used. Only if psi is not provided}
}
\value{
A list containing the value of the test statistic, the degrees of freedom, the p-values and the test decision
}
\description{
Computes the MCTP for clustered data
}
