% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{nm_gen}
\alias{nm_gen}
\title{Function generating sample- and cluster sizes}
\usage{
nm_gen(nn, n_i, m_ij, each_s = F, both_s = T, identical_s = T, identical_c = T)
}
\arguments{
\item{nn}{Number of groups/samples}

\item{n_i}{Typical sample sizes}

\item{m_ij}{Typical cluster size}

\item{each_s}{Logical: Does each sample contain a large/small cluster? Default: FALSE}

\item{both_s}{Logical: Does a sample contain both large/small clusters or only one of them? Default: TRUE}

\item{identical_s}{Logical: Identical Sample sizes? Default: TRUE}

\item{identical_c}{Logical: Identical Cluster sizes? Default: TRUE}
}
\value{
List containing the sample- and cluster sizes.
}
\description{
Function generating sample- and cluster sizes
}
