% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{q_wald}
\alias{q_wald}
\title{Computes the Wald-Test for clustered data}
\usage{
q_wald(n, data, cont, theta = NULL, psi = NULL, alpha = 0.05)
}
\arguments{
\item{n}{A vector containing the sample sizes}

\item{data}{The data, provided as a list of lists}

\item{cont}{A contrast matrix for the relative effects}

\item{theta}{A Vector for the group weights, defaults to unweighted estimator}

\item{psi}{A list of vectors with the cluster weights, defaults to unweighted estimator}

\item{alpha}{The significance level, defaults to 0.05}
}
\value{
A list containing the value of the test statistic, the degrees of freedom, the p-value and the test decision
}
\description{
Computes the Wald-Test for clustered data
}
