\name{Rankcluster-package}
\alias{Rankcluster-package}
\alias{Rankcluster}
\docType{package}
\title{
Model-based clustering for multivariate partial ranking data
}
\description{

This package proposes the first model-based clustering algorithm dedicated to multivariate partial ranking data. This is an extension of the Insertion Sorting Rank (ISR)[1] model for ranking data, which is a meaningful and effective model obtained by modelling the ranking generating process assumed to be a sorting algorithm.

The heterogeneity of the rank population is modelled by a mixture of ISR, whereas conditional independence assumption allows the exension to multivariate ranking. Maximum likelihood estimation process. 

}


\details{
\tabular{ll}{
Package: \tab Rankclust\cr
Type: \tab Package\cr
Version: \tab 0.9\cr
Date: \tab 2013-03-27\cr
License: \tab GPL (>=2) \cr
}
}

\author{
Maintainer: Grimonprez Quentin <quentin.grimonprez@inria.fr>
}

\references{
[1] C.Biernacki and J.Jacques (2012), A generative model for rank data based on sorting algorithm, Computational Statistics and Data Analysis, in press, DOI 10.1016/j.csda.2012.08.008

[2] J. Jacques, C. Biernacki (2012), Model-based clustering for multivariate partial ranking data 
}

\keyword{ package }

