% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{rel_eff}
\alias{rel_eff}
\title{Computes relative effects with respect to the reference distribution}
\usage{
rel_eff(data, theta = NULL, psi = NULL)
}
\arguments{
\item{data}{The data, provided as a list of lists}

\item{theta}{A vector of group weights, defaults to unweighted estimator}

\item{psi}{A list of vector-weights, defaults to unweighted estimator}
}
\value{
A vector of the relative effects
}
\description{
Computes relative effects with respect to the reference distribution
}
