% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultClass.R
\name{show,Output-method}
\alias{show,Output-method}
\alias{show,Rankclust-method}
\title{Show function.}
\usage{
\S4method{show}{Output}(object)

\S4method{show}{Rankclust}(object)
}
\arguments{
\item{object}{an object of class \link{Output-class} or \link{Rankclust-class}.}
}
\description{
This function shows the elements of a given object.
}
