% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sports}
\alias{sports}
\title{Rank data: sports}
\format{
A list containing:
\describe{
  \item{data}{a matrix containing 130 ranks of size 7 in ranking representation.

              The ranking representation r=(r_1,...,r_m) contains the  ranks assigned to the objects, and means that the ith object is in r_ith position.

              For example, if the ranking representation of a rank is (4,3,1,2,5), it means that judge ranks the first object in 4th position, second object in 3rd position, ...}

  \item{frequency}{a matrix with 123 differents ranks of size 7. In each row the first 7 columns correspond to one observed ranking and the last column contains the observation frequency.}
  \item{m}{ the size of the rankings (m=7).}
}
}
\source{
J.I. Marden. "Analyzing and modeling rank data, volume 64 of Monographs on Statistics and Applied Probability". Chapman & Hall, London, 1995.
}
\description{
This data set is due to Louis Roussos who asked 130 students at the
University of Illinois to rank seven sports according to their preference in participating:
 A = Baseball, B = Football, C = Basketball, D = Tennis, E = Cycling, F =
 Swimming, G = Jogging.
}
\examples{
data(sports)

}
\seealso{
Other datasets: 
\code{\link{APA}},
\code{\link{big4}},
\code{\link{eurovision}},
\code{\link{quiz}},
\code{\link{words}}
}
\concept{datasets}
\keyword{datasets}
