% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultClass.R
\name{[,Rankclust-method}
\alias{[,Rankclust-method}
\title{Getter method for rankclust output}
\usage{
\S4method{[}{Rankclust}(x, i, j, drop)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i}{the number of cluster of the element we want to extract or "bic", "icl", "ll"}

\item{j, drop}{not used}
}
\description{
Extract values of various
slots of the output from the function \code{\link{rankclust}}.
}
