% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultClass.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,Rankclust-method}
\title{summary function.}
\usage{
\S4method{summary}{Rankclust}(object, ...)
}
\arguments{
\item{object}{output object from \code{\link{rankclust}}.}

\item{...}{Not used.}
}
\description{
This function This function gives the summary of output from \code{rankclust}.
}
