% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultClass.R
\name{summary,Rankclust-method}
\alias{summary,Rankclust-method}
\title{Summary function.}
\usage{
\S4method{summary}{Rankclust}(object, ...)
}
\arguments{
\item{object}{output object from \code{\link{rankclust}}.}

\item{...}{Not used.}
}
\description{
This function gives the summary of an output from \code{rankclust}.
}
