% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{big4}
\alias{big4}
\title{rank data : big4}
\format{A list containing :
\describe{
  \item{data}{A matrix of size 21*8 containing the 21 Premier League seasons. Each row corresponding to one ranking (ranking representation).

             The ranking representation r=(r_1,...,r_m) contains the  ranks assigned to the objects, and means that the ith object is in r_ith position.

              For example, if the ranking representation of a rank is (4,3,1,2,5), it means that judge ranks the first object in 4th position, second object in 3rd position, ...}
  \item{frequency}{matrix of size 21*9. Each row corresponds to one of the 21 different observed rankings, and the last column contains the observation frequency.}
  \item{m}{the size of the rankings (m=c(4,4) ).}
}}
\source{
http://en.wikipedia.org/wiki/Big_Four_%28English_football%29#.22Big_Four.22_dominance_.282003.E2.80.9304_to_2008.E2.80.9309.29

http://www.uefa.com/memberassociations/uefarankings/club/index.html
}
\description{
This dataset is composed of the rankings (in ranking notation) of the "Big Four" English football teams (A: Manchester, B: Liverpool, C: Arsenal, D: Chelsea) to the English Championship (Premier League) and according to the UEFA coefficients (statistics used in Europe for ranking and seeding teams in international competitions), from 1993 to 2013.

In 2000-2001, Arsenal and Chelsea had the same UEFA coefficient and then are tied. UEFA ranking is (1, 4, 2, 2) for 2000-2001, what means that Manchester United is the first, Liverpool is the last, and the two intermediate positions are for Arsenal and Chelsea in an unknown order.

In 2009-2010, Liverpool and Arsenal have also the same UEFA coefficient, the ranking is  (1, 2, 2, 4).
}
\examples{
data(big4)
}
\keyword{datasets}

