% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{f_psi}
\alias{f_psi}
\title{Empirical Distribution function with cluster weights psi}
\usage{
f_psi(x, i, data, psi = NULL)
}
\arguments{
\item{x}{A scalar or vector to be checked as F(x)}

\item{i}{The group/sample index}

\item{data}{The data, provided as a list of lists}

\item{psi}{A vector of weights, defaults to unweighted estimator}
}
\value{
Value(s) of the empirical distribution function
}
\description{
Empirical Distribution function with cluster weights psi
}
