% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{[}
\alias{[}
\alias{[,Rankclust-method}
\title{Getter method for rankclust output}
\usage{
\S4method{[}{Rankclust}(x, i)
}
\arguments{
\item{x}{object from which to extract element(s) or in which to replace element(s).}

\item{i}{the number of cluster of the element we want to extract.}
}
\description{
This is overloading of square braces to extract values of various
slots of the output from the function \code{\link{rankclust}}.
}

