\name{[}
\alias{[}
\alias{[,Rankclust-method}
\title{Getter method for rankclust output}
\arguments{
  \item{x}{object from which to extract element(s) or in
  which to replace element(s).}

  \item{i}{the number of cluster of the element we want to
  extract.}
}
\description{
  This is overloading of square braces to extract values of
  various slots of the output from the function
  \code{\link{rankclust}}.
}
