\name{Rankclust-class}
\alias{Rankclust-class}
\title{Constructor of Rankclust class}
\description{
  This class contains results of rankclust function.
}
\details{
  \describe{ \item{K}{vector containing the number of
  clusters.} \item{data}{data used in algorithm.}
  \item{criterion}{criterion defined in rankclust function
  to select the best result.} \item{convergence}{if 0, no
  convergence, no result available in results.}
  \item{results}{a list of the same length than K,
  containing Output objects.} }
}

