\name{rankCluster-package}
\alias{rankCluster-package}
\alias{rankCluster}
\docType{package}
\title{
\packageTitle{rankCluster}
}
\description{
\packageDescription{rankCluster}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rankCluster}
\packageIndices{rankCluster}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{rankCluster}

Maintainer: \packageMaintainer{rankCluster}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
