\name{rankCluster-package}
\alias{rankCluster-package}
\alias{rankCluster}
\docType{package}
\title{
\packageTitle{rankCluster}
}
\description{
\packageDescription{rankCluster}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rankCluster}
\packageIndices{rankCluster}
}
\author{
\packageAuthor{rankCluster}

Maintainer: \packageMaintainer{rankCluster}
}
\references{
}

\keyword{ package }
\seealso{

}
\examples{
}
