% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultClass.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,Output-method}
\alias{show,Rankclust-method}
\title{show function.}
\usage{
\S4method{show}{Output}(object)

\S4method{show}{Rankclust}(object)
}
\arguments{
\item{object}{an object of class Output or Rankclust.}
}
\description{
This function shows the elements of a given object.
}

