\docType{methods}
\name{show}
\alias{show}
\alias{show,Output-method}
\alias{show,Rankclust-method}
\title{show function.}
\usage{
\S4method{show}{Output}(object)

\S4method{show}{Rankclust}(object)
}
\arguments{
  \item{object}{of class Output.}

  \item{object}{output object from
  \code{\link{rankclust}}.}
}
\description{
This function shows the elements of an object of class
Output.

This function shows the output from \code{rankclust}.
}

