% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.R
\name{sigma_est}
\alias{sigma_est}
\title{Computes the Variance-Covariance-Matrix of the relative effects}
\usage{
sigma_est(n, data, theta = NULL, psi = NULL)
}
\arguments{
\item{n}{A vector of sample sizes for each group}

\item{data}{The data, provided as a list of lists}

\item{theta}{A vector containing the group weights, defaults to unweighted estimator}

\item{psi}{A list of vector weights for the clusters, defaults to unweighted estimator}
}
\value{
A Variance-Covariance-Matrix
}
\description{
Computes the Variance-Covariance-Matrix of the relative effects
}
