\docType{data}
\name{sports}
\alias{sports}
\title{rank data : sports}
\format{A list containing :
\describe{
\item{data}{a matrix containing 130 ranks of size 7.}

\item{frequency}{a matrix with 123 differents ranks of size 7. In each row the first 7 columns correspond to one observed ranking and the last column contains the observation frequency.}
\item{m}{ the size of the rankings (m=7).}
}

}
\source{
  J.I. Marden. "Analyzing and modeling rank data, volume 64 of Monographs on
Statistics and Applied Probability". Chapman & Hall, London, 1995.

}
\description{
This data set is due to Louis Roussos who asked 130 students at the
University of Illinois to rank seven sports according to their preference in participating:
1 = Baseball, 2 = Football, 3 = Basketball, 4 = Tennis, 5 = Cycling, 6 =
Swimming, 7 = Jogging.

	
}
\examples{
data(sports)
}
\keyword{datasets}

