\docType{data}
\name{sports}
\alias{sports}
\title{rank data : sports}
\format{A list containing :

-data: matrix containinf 130 ranks of size 7

-frequency: matrix with 123 differents ranks of size 7. A row contains 1 rank of size 7 and the frequency.

-m: vector with the size of the rank for each dimension.}
\source{
  J.I. Marden. "Analyzing and modeling rank data, volume 64 of Monographs on
Statistics and Applied Probability". Chapman & Hall, London, 1995.

}
\description{
This data set is due to Louis Roussos who asked 130 students at the
University of Illinois to rank seven sports according to their preference in participating:
1 = Baseball, 2 = Football, 3 = Basketball, 4 = Tennis, 5 = Cycling, 6 =
Swimming, 7 = Jogging.

	
}
\examples{
data(sports)
}
\keyword{datasets}

