% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unfrequence}
\alias{unfrequence}
\title{Convert data}
\usage{
unfrequence(data)
}
\arguments{
\item{data}{a matrix containing rankings and observation frequency.}
}
\value{
a matrix containing all the rankings.
}
\description{
This function takes in input a matrix in which the m first columns are the different observed ranks and the last column contains the observation frequency, and returns a matrix containing all the ranks (ranks with frequency>1 are repeated).
}
\examples{
data(quiz)
Y=unfrequence(quiz$frequency)
Y
}

